+++ src/components/DiaryProductsListItem/DiaryProductsListItem.js
