+++ src/pages/DiaryPage.jsx
@@ -1,8 +1,8 @@
-import DiaryAddProductForm from "components/DiaryAddProductForm/DiaryAddProductForm";
-import DiaryProductsList from "components/DiaryProductsList/DiaryProductsList";
+import DiaryAddProductForm from '../components/DiaryAddProductForm/DiaryAddProductForm'
+import DiaryProductsList from "../components/DiaryProductsList/DiaryProductsList";
 
 export default function DiaryPage() {
-  return <>DiaryPage
+  return <>
     <DiaryAddProductForm />
     <DiaryProductsList />
   </>;
