+++ src/sass/main.scss
@@ -2,4 +2,5 @@
 @import "./base/global-styles";
 @import "./base/BaseStyles.scss";
 @import "./components/header";
+@import "../../src/components/DiaryAddProductForm/DiaryAddProductForm.module.scss"
